requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.0.15
langchain==0.0.350
pydantic==2.5.2
sqlalchemy==2.0.23
google-generativeai==0.3.1
python-multipart==0.0.6
